tarball: run container with host's /tmp
checkin.sh will build the tarball in the container's /tmp. Given it will
essentially be backed by a file somewhere, there is an expectation that
disk access will be slower than if we were running it on a proper
in-memory fs. The simplest answer might be to just pass our /tmp to the
container.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>